Renamed Twitch vue file and fixed dropdown location
