+++ src/components/widgets/index.ts
@@ -1,7 +1,7 @@
 export { default as Base } from "./Base.vue";
 export { default as Pauze } from "./Pauze.vue";
 export { default as VIA } from "./VIA.vue";
-export { default as Twitch } from "./TwitchFollow.vue";
+export { default as Twitch } from "./Twitch.vue";
 export { default as Spotify } from "./Spotify.vue";
 export { default as Battery } from "./BatteryWidget.vue";
 export { default as Dummy } from "./Dummy.vue";
